Posts/11_post_Fractal_GAN/requirements.txt
absl-py==2.0.0
altair==5.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bleach==6.1.0
blinker==1.6.2
bokeh==3.3.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
EbookLib==0.18
etils==1.5.2
executing==2.0.1
filelock==3.12.4
Flask==2.3.3
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.47.2
formulaic==1.0.1
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
interface-meta==1.3.0
ipython==8.21.0
iso8601==2.1.0
itsdangerous==2.1.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
linearmodels==5.4
linkify-it-py==2.0.2
live==0.0.3
llvmlite==0.42.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.3
packaging==23.2
pandas==2.2.0
panel==1.3.7
param==2.0.2
parso==0.8.3
patsy==0.5.6
pbr==5.11.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pyd==0.13.1
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyhdfe==0.2.0
pymobi==0.1.3
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
pyviz_comms==3.0.1
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
qiskit==0.44.1
qiskit-terra==0.25.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
rustworkx==0.13.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
serial==0.0.97
setuptools-scm==8.0.4
six==1.16.0
sklearn==0.0.post9
smmap==5.0.1
stack-data==0.6.3
statsmodels==0.14.1
stevedore==5.1.0
streamlit==1.31.1
symengine==0.9.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
tensorflowjs==4.14.0
tensorstore==0.1.45
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.2.1
torch-vision==0.1.6.dev0
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions @ file:///private/tmp/python-typing-extensions-20231212-4978-cvkltg/typing_extensions-4.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.24.5
Wave==0.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.7
word2number==1.1
wrapt==1.14.1
wurlitzer==3.0.3
xyzservices==2023.10.1
zipp==3.17.0
